# Conflicts:
#	packages/preview-middleware-client/src/flp/init.ts
#	packages/preview-middleware/README.md
#	packages/preview-middleware/src/base/flp.ts
#	packages/preview-middleware/test/unit/base/__snapshots__/flp.test.ts.snap

+++ packages/ui5-application-writer/src/options.ts
@@ -89,10 +89,13 @@ export async function enableTypescript(input: FeatureInput, keepOldComponent: bo
         ui5Config.addCustomTasks([ui5TSSupport.task]);
     });
     const compPath = join(input.basePath, 'webapp/Component.js');
-    if (keepOldComponent) {
-        input.fs.move(compPath, `${compPath}.old`);
-    } else {
-        input.fs.delete(compPath);
+    // Component.js may not exist if the project already uses TypeScript
+    if (input.fs.exists(compPath)) {
+        if (keepOldComponent) {
+            input.fs.move(compPath, `${compPath}.old`);
+        } else {
+            input.fs.delete(compPath);
+        }
     }
 }
